requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
gspread==6.1.2
google-auth==2.29.0
supabase==2.4.6
python-dotenv==1.0.1
